- docs: cover logging
