+++ engines/experiment/examples/plugins/hardcoded/go.mod
@@ -2,6 +2,9 @@ module github.com/gojek/turing/engines/experiment/examples/plugins/hardcoded
 
 go 1.14
 
-require github.com/gojek/turing/engines/experiment v1.0.0
+require (
+	github.com/gojek/turing/engines/experiment v1.0.0
+	github.com/hashicorp/go-hclog v0.16.0
+)
 
 replace github.com/gojek/turing/engines/experiment => ../../../
