+++ engines/experiment/log/log.go
@@ -1,5 +1,11 @@
 package log
 
+import (
+	"errors"
+	"github.com/hashicorp/go-hclog"
+	"go.uber.org/zap"
+)
+
 var global = newZapLogger()
 
 // Logger interface defines the common set of logging methods.
@@ -28,8 +34,21 @@ type Logger interface {
 	SetLevel(lvl string)
 }
 
-func SetGlobalLogger(l Logger) {
-	global = l
+func SetGlobalLogger(l interface{}) {
+	switch l.(type) {
+	case Logger:
+		global = l.(Logger)
+	case hclog.Logger:
+		global = &hcLogger{
+			Logger: l.(hclog.Logger),
+		}
+	case *zap.SugaredLogger:
+		global = &zapLogger{
+			SugaredLogger: l.(*zap.SugaredLogger),
+		}
+	default:
+		panic(errors.New("unsupported logger type"))
+	}
 }
 
 func Glob() Logger {
